backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
dashscope>=1.23.9
python-dotenv==1.0.0
